Moved routes to its own folder and deleted src2
