+++ src/routes/authRoutes.ts
