+++ src/server.ts
@@ -1,8 +1,8 @@
-import authRoutes from 'authRoutes';
 import cookieParser from 'cookie-parser';
 import cors from 'cors';
 import dotenv from 'dotenv';
 import express, { Express, NextFunction, Request, Response } from 'express';
+import authRoutes from 'routes/authRoutes';
 
 dotenv.config();
