--- src2/server.js
@@ -1,80 +1,0 @@
-const express = require('express');
-const cors = require('cors');
-const cookieParser = require('cookie-parser');
-require('dotenv').config();
-
-const authRoutes = require('./routes/authRoutes');
-
-const app = express();
-
-const corsOptions = {
-  origin: (origin, callback) => {
-    const allowedOrigins = [
-      process.env.FRONTEND_URL,
-      process.env.FRONTEND_URL_DEV,
-    ];
-
-    if (allowedOrigins.includes(origin) || !origin) {
-      callback(null, true);
-    } else {
-      callback(new Error('Not allowed by CORS'));
-    }
-  },
-  methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
-  allowedHeaders: ['Content-Type', 'Authorization'],
-  credentials: true,
-  maxAge: 86400,
-};
-
-app.use(cors(corsOptions));
-
-if (process.env.NODE_ENV !== 'production') {
-  console.log('CORS Configuration:', {
-    origin: process.env.FRONTEND_URL,
-    credentials: true,
-  });
-}
-
-app.use(cookieParser());
-app.use(express.json());
-
-app.use((req, res, next) => {
-  req.url = req.url.replace(/\/+/g, '/');
-  next();
-});
-
-app.use('/auth', authRoutes);
-
-app.get('/health', (req, res) => {
-  res.status(200).json({ status: 'ok' });
-});
-
-// eslint-disable-next-line no-unused-vars
-app.use((err, req, res, next) => {
-  console.error('Error details:', {
-    message: err.message,
-    stack: err.stack,
-    status: err.status || 500,
-  });
-
-  res.status(err.status || 500).json({
-    error:
-      process.env.NODE_ENV === 'production'
-        ? 'Internal Server Error'
-        : err.message,
-  });
-});
-
-if (process.env.NODE_ENV !== 'production') {
-  console.log('CORS Configuration:', {
-    allowedOrigins: [process.env.FRONTEND_URL, process.env.FRONTEND_URL_DEV],
-    credentials: true,
-  });
-}
-
-const PORT = process.env.PORT || 3000;
-app.listen(PORT, () => {
-  console.log(`Server running on port ${PORT}`);
-  console.log(`Environment: ${process.env.NODE_ENV || 'development'}`);
-  console.log(`Frontend URL: ${process.env.FRONTEND_URL}`);
-});
